Convert `config.py` into `config` package
- Split into `config/__init__.py` and `config/utils.py`
